Implemented & tested

+++ Editor/MaterialView.cpp
@@ -142,9 +142,9 @@ namespace ToolKit
 
         ImGui::SameLine();
         const ImVec2 viewportSize = ImVec2(ImGui::GetContentRegionAvail().x -
-                                               iconSize.x - 5.0 * spacing.x,
+                                               iconSize.x - 5.0f * spacing.x,
                                            150.0f);
-        m_viewport->ResizeWindow(viewportSize.x, viewportSize.y);
+        m_viewport->ResizeWindow((uint) viewportSize.x, (uint) viewportSize.y);
         m_viewport->Update(g_app->GetDeltaTime());
         m_viewport->Show();
         ImGui::SameLine();
